EMP-21046 - create single app navigation reload func
