+++ SDKSampleApp/ViewControllers/LoginViewController.swift
@@ -181,12 +181,7 @@ extension LoginViewController {
                         StorageProvider.store(environment: self?.environment)
                         StorageProvider.store(sessionToken: credentials.sessionToken)
                         
-                        let navigationController = MainNavigationController()
-                        let castContainerVC = GCKCastContext.sharedInstance().createCastContainerController(for: navigationController)
-                          as GCKUICastContainerViewController
-                        castContainerVC.miniMediaControlsItemEnabled = true
-                        UIApplication.shared.keyWindow?.rootViewController = castContainerVC
-
+                        reloadAppNavigation()
                     }
                 }
             
@@ -212,12 +207,7 @@ extension LoginViewController {
                     StorageProvider.store(environment: self?.environment)
                     StorageProvider.store(sessionToken: credentials.sessionToken)
                     
-                    let navigationController = MainNavigationController()
-                    let castContainerVC = GCKCastContext.sharedInstance().createCastContainerController(for: navigationController)
-                      as GCKUICastContainerViewController
-                    castContainerVC.miniMediaControlsItemEnabled = true
-                    UIApplication.shared.keyWindow?.rootViewController = castContainerVC
-
+                    reloadAppNavigation()
                 }
         }
     }
